Move Operation to its final place
